moving addForm to new component and refactoring
