+++ src/views/List.jsx
@@ -3,8 +3,8 @@ import { ListItem } from '../components';
 import { useNavigate } from 'react-router-dom';
 import { comparePurchaseUrgency } from '../api/firebase';
 
-export function List({ data, listPath }) {
-	console.log(data.data, listPath);
+export function List({ data, listPath, listName }) {
+	console.log(data.data, listPath, listName);
 	const [input, setInput] = useState('');
 
 	const navigate = useNavigate();
